backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
llama-index
llama-index-embeddings-google-genai
llama-index-vector-stores-supabase
python-dotenv